+++ src/pageEditor/slices/editorThunks.ts
@@ -18,10 +18,27 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import { EditorRootState } from "@/pageEditor/pageEditorTypes";
 import { produce } from "immer";
-import { selectActiveElement } from "@/pageEditor/slices/editorSelectors";
+import {
+  selectActiveElement,
+  selectNotDeletedElements,
+  selectNotDeletedExtensions,
+} from "@/pageEditor/slices/editorSelectors";
 import { uuidv4 } from "@/types/helpers";
 import { normalizePipelineForEditor } from "@/pageEditor/extensionPoints/pipelineMapping";
 import { actions } from "@/pageEditor/slices/editorSlice";
+import { UUID } from "@/core";
+import { ExtensionsRootState } from "@/store/extensionsTypes";
+import {
+  checkAvailable,
+  getInstalledExtensionPoints,
+} from "@/contentScript/messenger/api";
+import { getCurrentURL, thisTab } from "@/pageEditor/utils";
+import { resolveDefinitions } from "@/registry/internal";
+import { QuickBarExtensionPoint } from "@/extensionPoints/quickBarExtension";
+import { testMatchPatterns } from "@/blocks/available";
+import { BaseExtensionPointState } from "@/pageEditor/extensionPoints/elementConfig";
+import { isQuickBarExtensionPoint } from "@/pageEditor/extensionPoints/formStateTypes";
+import { compact, uniq } from "lodash";
 
 export const cloneActiveExtension = createAsyncThunk<
   void,
@@ -44,4 +61,97 @@ export const cloneActiveExtension = createAsyncThunk<
   );
   // Add the cloned extension
   thunkAPI.dispatch(actions.addElement(newElement));
+});
+
+export type AvailableInstalled = {
+  availableInstalledIds: UUID[];
+  unavailableCount: number;
+};
+
+export const checkAvailableInstalledExtensions = createAsyncThunk<
+  AvailableInstalled,
+  void,
+  { state: EditorRootState & ExtensionsRootState }
+>("editor/checkAvailableInstalledExtensions", async (arg, thunkAPI) => {
+  const extensions = selectNotDeletedExtensions(thunkAPI.getState());
+  const installedExtensionPoints = new Map(
+    // eslint-disable-next-line unicorn/no-await-expression-member
+    (await getInstalledExtensionPoints(thisTab)).map((extensionPoint) => [
+      extensionPoint.id,
+      extensionPoint,
+    ])
+  );
+  const resolved = await Promise.all(
+    extensions.map(async (extension) => resolveDefinitions(extension))
+  );
+  const tabUrl = await getCurrentURL();
+  const availableExtensionPointIds = resolved
+    .filter((x) => {
+      const extensionPoint = installedExtensionPoints.get(x.extensionPointId);
+      // Not installed means not available
+      if (extensionPoint == null) {
+        return false;
+      }
+
+      // QuickBar is installed on every page, need to filter by the documentUrlPatterns
+      if (QuickBarExtensionPoint.isQuickBarExtensionPoint(extensionPoint)) {
+        return testMatchPatterns(extensionPoint.documentUrlPatterns, tabUrl);
+      }
+
+      return true;
+    })
+    .map((x) => x.id);
+
+  const availableInstalledIds = extensions
+    .filter((x) => availableExtensionPointIds.includes(x.id))
+    .map((x) => x.id);
+  const unavailableCount = extensions.length - availableInstalledIds.length;
+
+  return { availableInstalledIds, unavailableCount };
+});
+
+async function isElementAvailable(
+  tabUrl: string,
+  elementExtensionPoint: BaseExtensionPointState
+): Promise<boolean> {
+  if (isQuickBarExtensionPoint(elementExtensionPoint)) {
+    return testMatchPatterns(
+      elementExtensionPoint.definition.documentUrlPatterns,
+      tabUrl
+    );
+  }
+
+  return checkAvailable(thisTab, elementExtensionPoint.definition.isAvailable);
+}
+
+export const checkAvailableDynamicElements = createAsyncThunk<
+  { availableDynamicIds: UUID[] },
+  void,
+  { state: EditorRootState }
+>("editor/checkAvailableDynamicElements", async (arg, thunkAPI) => {
+  const elements = selectNotDeletedElements(thunkAPI.getState());
+  const tabUrl = await getCurrentURL();
+  const availableElementIds = await Promise.all(
+    elements.map(async ({ uuid, extensionPoint: elementExtensionPoint }) => {
+      const isAvailable = await isElementAvailable(
+        tabUrl,
+        elementExtensionPoint
+      );
+
+      return isAvailable ? uuid : null;
+    })
+  );
+
+  return { availableDynamicIds: uniq(compact(availableElementIds)) };
+});
+
+export const checkActiveElementAvailability = createAsyncThunk<
+  { isAvailable: boolean },
+  void,
+  { state: EditorRootState }
+>("editor/checkDynamicElementAvailability", async (arg, thunkAPI) => {
+  const tabUrl = await getCurrentURL();
+  const element = selectActiveElement(thunkAPI.getState());
+  const isAvailable = await isElementAvailable(tabUrl, element.extensionPoint);
+  return { isAvailable };
 });
